[fix] Adds background flag to object detection scripts
This commit adds a `background` flag to object detection scripts.
When this flag is added, the script will output the background class
as index 0 and all other class indices are one greater than the class
index in the Visual Genome category mapping. This commit also updates
the non-maximum suppression calculation and argmax for object class
prediction to use the starting index controlled by the `background`
flag. When the `background` flag is active, the starting index is
0 to include the background class. Otherwise, it is 1.