[fix] Fix extract_features_vmb.py and extract_features.py to predict correct object classes (#127)
* [fix] Fix extract_features_vmb.py and extract_features.py to predict correct object classes

This commit excludes the background class (index 0 of the scores vector) when computing
the object class. This prevents the model from predicting the background class when
the score of the background class is larger than the score computed using non-maximum
suppression over non-background classes. This change also has the added benefit of
aligning the object indices predicted by the model with the object mapping for Visual
Genome (https://dl.fbaipublicfiles.com/pythia/data/visual_genome_categories.json).

* [fix] Adds background flag to object detection scripts

This commit adds a `background` flag to object detection scripts.
When this flag is added, the script will output the background class
as index 0 and all other class indices are one greater than the class
index in the Visual Genome category mapping. This commit also updates
the non-maximum suppression calculation and argmax for object class
prediction to use the starting index controlled by the `background`
flag. When the `background` flag is active, the starting index is
0 to include the background class. Otherwise, it is 1.