[refactor] Refactor Vilbert and Visualbert (#67)
Summary:
Refactor of our BERT based models. This removes the use of sample list from our `*Pretraining` or `*Classification` heads and these modules can be called directly with tensors or strings inputs. This will help to make these modules scriptable. Enabling scripting for these models will come in next set of PRs.

This PR also consolidates all classification heads for different datasets. `training_head_type` is set to `classification` for all along with a `num_labels` configuration. For nlvr2 we are keeping the `training_head_type` as `nlvr2` as we need to specially modify the `hidden_size` for that dataset.

Tested with current visual bert/vilbert models as well as loading old models.
Pull Request resolved: https://github.com/fairinternal/pythia-internal/pull/67

Reviewed By: apsdehal

Differential Revision: D21272728

Pulled By: vedanuj

fbshipit-source-id: 715af8be62caa1e4e10f84a63eb45499f30a6362