[enhancement,refactor,fix] M4C Captioner, TextCaps related changes (#113)
Summary:
- Fix test reporter according to new API, distributed sampler
- MultiDataset distributed sampler will be loaded now
- Refactor M4C TextCaps to TextCaps
- Remove the confusion between COCO and TextCaps by introducing
annotation_style config parameter that will allow to load other datasets
than default dataset. In long term, we want datasets to be able to
specify the dataset class that will be used to build the dataset
- Allow `zoo_requirements` in the model zoo as well
- Upload detectron weights as well so that they can be included as
requirement
- Update all models to use zoo based detectron weights
- Add dataset zoo for textcaps, coco-m4c version and coco-ocr_en version
Pull Request resolved: https://github.com/fairinternal/pythia-internal/pull/113

Test Plan:
Same as the previous PR, you can use my data dir to test it out.
I have tested all of the variations that come with m4c_textcaps so we
should be good to go

Reviewed By: vedanuj

Differential Revision: D21421244

Pulled By: apsdehal

fbshipit-source-id: 5a89ee516f7e4f4cac68f7a977b893ace9af4a92