[fix,refactor] Add mmf_cli cmds, fix,speed up hm_convert, pip package (#262)
Summary:
- Python zipfile module uses Python interpretation to unzip files which
are encrypted; so improve performance we first use unzip command if
present otherwise fall back to zipfile implementation
- Bump the version number
- Add a new folder mmf_cli which will contain all cli commands in one
place but usually should symlink to tools folder
- Add proper __init__.py files for folder which were not getting
imported properly
- get_absolute_path now actually returns absolute path if the path is
not uri instead of directly returning the path. This fixes cases where
USER_DIR is set to "."
- If the folder is not git repository such as package folder git repo
creation in checkpoint used to throw error. Now, we catch that error and
just pass on it in case it is not a git repo
- Manifest file has been removed as package_data handles its uses
- Symlinks to packages that need to be included in mmf wheel have been
added in package subfolder so that they are easily includable as per
setuptools recommendation
- Automatically discovers the package data now
- Updates checksum for visual_bert coco
- Fix predict command for single GPU
Pull Request resolved: https://github.com/facebookresearch/mmf/pull/262

Test Plan: Tested first by publishing on test pypi

Reviewed By: vedanuj

Differential Revision: D21653287

Pulled By: apsdehal

fbshipit-source-id: 06b3b9d818b992b6f015a10d1b0215c9e07f3405