[refactor] Gathering of distributed tensors based on number of dimensions (#384)
Summary:
Pull Request resolved: https://github.com/facebookresearch/mmf/pull/384

Refactor to automatically check the dimensions and apply the appropriate reshaping before gathering distributed tensors. To make writing custom reports more convenient when new keys are added. �

Did not change any of the actual logic. Makes the reshaping explicit depending on the number of the dimensions. Gathering distributed tensors increases the number of dimensions by 1 and we reshape accordingly.

Will fail more gracefully if the tensor shape before `gather_tensor` is 0 or greater than 2.

Reviewed By: vedanuj, apsdehal

Differential Revision: D22299088

fbshipit-source-id: dbe71227a147bacf426fe3b6cc5c7eea463f8c7e