[feat] check unused model parameters in optimizer
Add optimizer option on whether to allow some of the model's parameters
not to be used by the optimizer. Default is false to guard against
missing parameters due to implementation errors in a model's
`get_optimizer_parameters`. Previously `get_optimizer_parameters` is
error-prong as the users need to remember to update
`get_optimizer_parameters` whenever adding a new submodule, and the
error is hard to spot when the users forget to do so.

Test plan: tested locally w/ VisualBERT, against correctly and
incorrectly implemented `get_optimizer_parameters`.