[feat] Synchronize device for batch based on model (#456)
Summary:
Pull Request resolved: https://github.com/facebookresearch/mmf/pull/456

Purpose of this diff is to add `to_device` method which is a helper utility method to move sample list to a proper device. In case the user is using custom dataset and forgot to move it to proper device, this changes will make sure that the batch is automatically moved to proper device at multiple stages of a training loop including making sure the batch model is receiving is on same device as the model's device.

The checks are added at the following stages:
- Before forward in training loop
- Before forward in prediction loop
- Inside BaseModel's forward

This diff also remove the warning about prepare_batch in the multi dataset loader as it is not needed anymore.

Reviewed By: vedanuj

Differential Revision: D22955817

fbshipit-source-id: 5cd69cff518480f75dd8e9eda14aca31e575c48b