[fix,feat] Add loss for in-batch cross entropy and allow "target" to be specified on the fly
Summary:
Previously, losses require "targets" in sample_list, and will silently return `{}` if "targets" are not present in `sample_list`, but this might not be the case for custom loss functions.

Self-supervised/metric-learning/retrieval based functions like in batch cross entropy (example provided) or alignment losses compute their targets on the fly within the batch, and leaving this flexible allows for different targets to be specified for different loss functions on the fly.

Differential Revision: D23269219

fbshipit-source-id: 43ec3b9a1f92cb788ed75c566f80cc8b139cb6ae