[refactor] Refactor MMF Transformer embeddings to support scriptability (#551)
Summary:
- Changes to support torchscript for mmf transformer derivatives
- Changes to `MMFTransformerEmbeddings` by modifying the different modality embeddings to `nn.ModuleList` to make it scriptable.
- Some changes to the forward function

Note: It is difficult to make MMF transformer scriptable as a whole since its base can be any type of transformer and making that scriptable is dependent on other libraries. The idea is to have derived models from MMF transformer that can be scripted. For example: We will have a version of MMF transformer that works with Bert, Roberta and XLMR that will be scriptable.

Pull Request resolved: https://github.com/facebookresearch/mmf/pull/551

Test Plan: - Test with hateful memes dataset

Reviewed By: apsdehal

Differential Revision: D23569270

Pulled By: vedanuj

fbshipit-source-id: 21d9ac80d912fcc65810a13f4facd0e828fcdd39