[feat] check unused model parameters in optimizer (#580)
Summary:
This PR adds optimizer option on whether to allow some of the model's parameters not to be used by the optimizer. Default is false to guard against missing parameters due to implementation errors in a model's `get_optimizer_parameters`. Previously `get_optimizer_parameters` is error-prong as the users need to remember to update `get_optimizer_parameters` whenever adding a new submodule, and the error is hard to spot when the users forget to do so.

Pull Request resolved: https://github.com/facebookresearch/mmf/pull/580

Test Plan:
tested locally w/ VisualBERT, against correctly and
incorrectly implemented `get_optimizer_parameters`.

Reviewed By: vedanuj

Differential Revision: D24201583

Pulled By: ronghanghu

fbshipit-source-id: 139f9837e7bc8f5276e33c12bee7f3061f3fa6a5