[feat] Add init dataclasses for mmbt and encoders (#565)
Summary:
Pull Request resolved: https://github.com/facebookresearch/mmf/pull/565

As step one of FAIM integration, we allow building our models from config so that the models are purely decoupled from configuration and users know what the model expects.

We first do this for the MMBT model.

- Adds configs for MMBT and respective encoders.
- Also adds from_params method to the BaseModel class so that args initialization is possible
- Updates build method to support passing of direct config object
- Add Config class to BaseModel as well and update typings
There is an issue with OmegaConf that doesn't let us use Union in structured configs. Take a look at https://github.com/omry/omegaconf/issues/144

Differential Revision: D23699688

fbshipit-source-id: 29f738d82c617c528212118ddbb60ee89bfca442