[feat] Add encoder registry and differentiate encoder factory (#628)
Summary:
Pull Request resolved: https://github.com/facebookresearch/mmf/pull/628

This diff aims to smoothen the process of adding new encoders to MMF. This is also specifically needed for FB internal encoders which are hard to registry while not being part of OSS without registry. This will also help with keeping sanity in MMFTransformer and generic modalities diff.

This diff also converts the older feature encoders to the factory which is what they technically are. This backwards compatible change as the build api still looks same. Though if someone is directly importing the encoder, this will break their code. But this refactor is much needed at this time.

We also add `build_encoder` function which handles both structured config and normal config patterns. We hope to make it more streamlined in future when we make efforts towards Hydra integration.

Reviewed By: vedanuj

Differential Revision: D24300517

fbshipit-source-id: 87c34648984c4e3f5417dfa6e3d965acc2404bcf