[fix, test] Fix recall metric bug, add test (#746)
Summary:
- Fix infinite recursion error https://github.com/facebookresearch/mmf/issues/725
- Add back mistakenly removed code
- Add test for recall metric at 1, 5, and 10

This is a response to bug https://github.com/facebookresearch/mmf/issues/725

I believe `process_ranks` may have been accidentally deleted in this commit so I added it back:
https://github.com/facebookresearch/mmf/commit/6b6f4fc221c46f8c55d55a06e110f3d8f9ca2417

I also added a testing file, which basically ranks all of the expected values in increasing order (so the last candidate is always the chosen one).

Both the expected values and predicted values are (10, 100) tensors. As the index of top level array increases, I alter the final value to be decreasing by 2, so the chosen candidates are 198 (maximum value 99*2), 191, 189, 187, ...

I decrease it in this way so that each of the candidates become out of the bounds for the increasing Recall Metrics. The rankings are `tensor([ 1,  4,  5,  6,  7,  8,  9, 10, 11, 12])`

This makes the results varying for each metric:

`RecallAt1` --> 1/10 = .1
`RecallAt5` --> 3/10 = .3
`RecallAt10` --> 8/10 = .8

I would love for people to take a look at `metrics.py` line `462` (`process_ranks`)
There are print statements in here from the previous commit, but I do not see print statements anywhere else in the modules, should I delete these, leave them, do we have a different logging style? Thanks!

Pull Request resolved: https://github.com/facebookresearch/mmf/pull/746

Reviewed By: vedanuj

Differential Revision: D26049706

Pulled By: brettallenyo

fbshipit-source-id: 03546c76f9aed3a1c26b318eefb00fbb05d1297a