[mmf][site] Upgrade Docusaurus
Summary:
To use FbInternalOnly, you need to be on Docusaurus 61+, this upgrades it.

It also fixes a typo FBInternalOnly -> FbInternalOnly.
MDX currently silently (with a warning) accepts components that haven't been imported, so this wasn't raising an error but wasn't doing anything either.

Also fixes a few broken links (the new version of docusaurus checks this for you and enforce them).

Reviewed By: apsdehal

Differential Revision: D26510193

fbshipit-source-id: 81871d3151edce27a2281cbcaaa7e694309205a2