[feat] add visualization API and notebook (#766)
Summary:
Add a visualization API for the frcnn
features, which takes in image and feature
paths and uses the huggingface visualization
tools to show the image.

Also added a jupyter notebook that will display
the images accordingly if the features and images
are in the correct place.

Pull Request resolved: https://github.com/facebookresearch/mmf/pull/766

Reviewed By: vedanuj

Differential Revision: D26621835

Pulled By: brettallenyo

fbshipit-source-id: a31e1a3da499c4044d6cf275b281ab5e4b0b720d