[refactor] Move test_reporter_config to evaluation.reporter (#838)
Summary:
Pull Request resolved: https://github.com/facebookresearch/mmf/pull/838

Test reporter config as opposed to original design should be a top level config node as it is not specific to one dataset but encompasses all datasets in one go. This diff refactors that to be a top level config node so that it is not per dataset config which conflicts with overall definition of test reporter.

Reviewed By: madian9

Differential Revision: D27301920

fbshipit-source-id: 6a990cd23c837fb915398a5930b9e348ff4cd626