[mmf] validate batch sizes are same + debugging improvements (#902)
Summary:
[feat]
Pull Request resolved: https://github.com/facebookresearch/mmf/pull/902

Before adding batch to reporter, sync with all workers and validate all have same batch size. If not, skip the batch. Reports in the end how many were skipped.

This should add quite negligible overhead. And once the problem itself is fixed (probably by constructing the dataloader with 'drop_incomplete' -- but i am not sure where to add this), this could be removed -- but perhaps better to just keep it.

In addition:
- remove master's logging to a file and instead log to stderr
- add CompleteInTimeOrDie() so that we get better traces than NCCL deadlocks cause.
- some tiny changes.

Reviewed By: apsdehal

Differential Revision: D27954629

fbshipit-source-id: cf000586772f4510b717153ec66e1f9c6c269ac3