[feat] add config option to retry previous batch when NaN loss happens
Add config option `training.try_previous_batch_on_nan_losses`. When setting to True, if the current batch results in NaNs in the losses, try forwarding on the previous batch in this iteration. This could be useful in avoiding occastional underflow or overflow when training.fp16 is activated. Mutually exclusive with training.exit_on_nan_losses

Test plan: tested locally in MMF training