[feat] Add metric support for dict return type
Summary:
Add metric support for classes that return a dict of metrics.
All metric dictionaries are flattened in the final metrics dict.
Add example f1_precision_recall metric class as an example,
and for unit testing. This new metric class returns a dict,
containing the f1, precision, and recall tensors as values
with string keys.