[feat] Add VinVL text tokenizer (#1152)
Summary:
Pull Request resolved: https://github.com/facebookresearch/mmf/pull/1152

Add text tokenizer which can do process
masked text for mlm and corrupt text for
VinVL 3-way contrastive loss.

Test Plan:
**Unit tests**

Tests for the construction of the processor and processor output.
We assert that the tokens are correct and contain the enhanced fields required for VinVL.

Reviewed By: apsdehal

Differential Revision: D32574737

Pulled By: Ryan-Qiyu-Jiang

fbshipit-source-id: da2c68cf5fa388cb31efe1ce2fab012e1eb8da73