docs: add social banner in support of Ukraine (#1207)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

![image](https://user-images.githubusercontent.com/12485205/156669684-0beec8b2-1d63-4eab-890a-0130bcd481c4.png)

Pull Request resolved: https://github.com/facebookresearch/mmf/pull/1207

Test Plan:
Imported from GitHub, without a `Test Plan:` line.

**Static Docs Preview: mmf**
|[Full Site](https://our.intern.facebook.com/intern/staticdocs/eph/D34628051/V2/mmf/)|

|**Modified Pages**|

Reviewed By: apsdehal

Differential Revision: D34628051

Pulled By: dmitryvinn-fb

fbshipit-source-id: 2099660b137cdd1bffa3188ba0dcec25865d627c