fix unexpected ValueError
When value is a numpy array, comparing a numpy array with None object will give another numpy array. Then a ValueError will be raised, says: ValueError: The truth value of an array with more than one element is ambiguous. Use a.any() or a.all().
For example, you can test this funtion with value = numpy.array([1,2,3]), and default=None.