Cleanup transformers3 mess
Summary:
transformers (as imported internally) has a convoluted and bloody history.

the current state is that "transformers3" exists alongside "transformers", and makes things messy for upgrades and rollouts, especially considering the hacks applied to "transformers3" historically.

this diff attempts to get closer to a cleaner state, by deleting "transformers3" after existing usage was migrated to "transformers" in the previous diff

Reviewed By: lisroach, diliop

Differential Revision: D68842833

fbshipit-source-id: 13b80b92ee1005f26e6a60907566bd5cc6167a86